+++ controllers/batch-deletion/api-requests.js
@@ -71,4 +71,28 @@ const getDeletionBatches = (req, res, next) => {
     });
 };
 
-module.exports = { getDeletionBatches };
+const sendFile = async (req, res, next) => {
+  const { fileContent, batchTitle } = req.body;
+  const targetRefIds = fileContent.split("\n").map((item) => item.trim());
+  try {
+    const response = await api(req, { adminApi: true }).post(
+      "/deletion-batches/",
+      {
+        json: {
+          name: batchTitle,
+          targetRefDomain: "domain",
+          targetRefIds,
+        },
+      }
+    );
+    // todo: React to certain responses
+    console.log(response);
+
+    res.status(200).send({ message: "File sent successfully" });
+  } catch (error) {
+    console.log("error: ", error.statusCode);
+    next(error);
+  }
+};
+
+module.exports = { getDeletionBatches, sendFile };
