+++ controllers/batch-deletion/batch-deletion.js
@@ -5,14 +5,12 @@
 const express = require("express");
 const router = express.Router();
 const authHelper = require("../../helpers/authentication");
-const { api } = require("../../api");
-const apiRequests = require("./fetch-data");
-const sendFile = require("./send-file");
+const apiRequests = require("./api-requests");
 
 router.use(authHelper.authChecker);
 
 router.get("/", apiRequests.getDeletionBatches);
 
-router.post("/create-batch-deletion-file",sendFile);
+router.post("/create-batch-deletion-file", apiRequests.sendFile);
 
 module.exports = router;
